requirements.txt
altgraph==0.16.1
attrs==19.1.0
auto-py-to-exe==2.5.2
backcall==0.1.0
bleach==3.1.0
bokeh==1.0.4
bottle==0.12.16
bottle-websocket==0.2.9
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cloudpickle==0.8.1
colorama==0.4.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
Eel==0.10.4
entrypoints==0.3
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.3.3
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.17.0
packaging==19.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.4.0
pefile==2018.8.8
pickleshare==0.7.5
Pillow==6.0.0
plotly==3.7.1
prometheus-client==0.7.0
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.3.1
PyInstaller==3.4
pyparsing==2.3.1
# pypiwin32==223
pyrsistent==0.14.11
python-dateutil==2.8.0
python-pptx==0.6.17
pytz==2018.9
# pywin32==224
# pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder-kernels==0.4.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.1
urllib5==5.0.0
wcwidth==0.1.7
webencodings==0.5.1
whichcraft==0.5.2
widgetsnbextension==3.4.2
xlrd==1.2.0
XlsxWriter==1.1.5
zope.interface==4.6.0
